requirements.txt
django
djangorestframework
django-allauth
django-cors-headers
